+++ src/python/pants/backend/python/rules/importable_python_sources_test.py
@@ -8,7 +8,7 @@ from pants.backend.python.rules.importable_python_sources import (
     rules as importable_python_sources_rules,
 )
 from pants.build_graph.address import Address
-from pants.core.targets import Files
+from pants.core.target_types import Files
 from pants.engine.selectors import Params
 from pants.engine.target import Sources, Target, Targets
 from pants.testutil.option.util import create_options_bootstrapper
