+++ src/python/pants/core/rule_utils/strip_source_roots.py
@@ -7,7 +7,7 @@ from pathlib import PurePath
 from typing import Optional, cast
 
 from pants.build_graph.address import Address
-from pants.core.targets import FilesSources
+from pants.core.target_types import FilesSources
 from pants.engine.fs import (
     EMPTY_SNAPSHOT,
     Digest,
